fix BLE and OTA errors

+++ extra_script.py
@@ -17,36 +17,38 @@ bin = project_config.get("common", "JsonBin")
 
 
 def after_build(source, target, env):
-    print("after_build !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!")
+    print("----------------------------------------------------------------------------------------------------------")
     firmware_path = str(source[0])
-    print(firmware_path)
-    firmware_name = basename(firmware_path)
-    #print(env.Dump())
-    
-    
     envName = env["PIOENV"]
-    print(envName)
     dstMainPath = env["PROJECT_DIR"] + "\\ota_updates\\" + envName + "\\"
     dstFilename = "firmware_v" + version + ".bin"
-    print(dstMainPath)
+    finalDstFullPath = dstMainPath + dstFilename
+    
+    if os.path.isfile(finalDstFullPath):
+        print ("This firmware version already exists ... increase the firmware_version or delete the files if you want to generate a new version.")
+    else:
+        print("Copying bin file...")
+        print(finalDstFullPath)
 
-    os.makedirs(os.path.dirname(dstMainPath), exist_ok=True)
-    copyfile(firmware_path , dstMainPath + dstFilename)
+        os.makedirs(os.path.dirname(dstMainPath), exist_ok=True)
+        copyfile(firmware_path , finalDstFullPath)
 
-    # Generating JSON version file
-    print("Generating JSON version file...")
-    payload = {}
-    payload['type'] = envName
-    payload['version'] = version
-    payload['host'] = host
-    payload['port'] = port
-    payload['bin'] = bin + envName + "/" + dstFilename
+        # Generating JSON version file
+        print("Generating JSON version file...")
+        payload = {}
+        payload['type'] = envName
+        payload['version'] = version
+        payload['host'] = host
+        payload['port'] = port
+        payload['bin'] = bin + envName + "/" + dstFilename
 
-    try:
-        with open(dstMainPath + "firmware.json", 'w') as outfile:
-            json.dump(payload, outfile)
-    except:
-        print("[ERROR] Error when generating the JSON version file...")
+        try:
+            with open(dstMainPath + "firmware.json", 'w') as outfile:
+                json.dump(payload, outfile)
+        except:
+            print("[ERROR] Error when generating the JSON version file...")
+
+    print("----------------------------------------------------------------------------------------------------------")
 
 
 env.AddPostAction("buildprog", after_build)
